apr_socket_sendfile:
Be accurate (and consistent accross unixes) with the returned number of bytes
written (note that it is possible for both bytes to be sent and an error to be
returned, due to headers and trailers handling).

Fix some returned errno vs apr_status_t in some error paths.

Fix BSDs nonblocking/busy retry on EAGAIN when nothing is written.
 


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1789257 13f79535-47bb-0310-9956-ffa450edef68