Win32: Create and use file mutex only for files opened with APR_FOPEN_XTHREAD.
There are some cases in the Windows implementation of the file I/O that
cause the unnecessary creation and acquisitions of the file mutex.
These cases include handling O_APPEND-style writes and the implementation
of the apr_file_buffer_set().

Creating and acquiring the file mutex is only required for files opened
with the APR_FOPEN_XTHREAD flag.  Otherwise, concurrent operations on
the same apr_file_t instance should be serialized by the user, and
the mutex is not required.

This patch tweaks the implementation to only create and acquire/release
the mutex for files opened with APR_FOPEN_XTHREAD.

* file_io/win32/open.c
  (apr_file_open, apr_os_file_put): Create the file mutex only
   with APR_FOPEN_XTHREAD.

* file_io/win32/buffer.c
  (apr_file_buffer_set): Use the file mutex only with APR_FOPEN_XTHREAD.

* file_io/win32/readwrite.c
  (apr_file_write): Use the file mutex only with APR_FOPEN_XTHREAD.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1808457 13f79535-47bb-0310-9956-ffa450edef68