Merge r1814239, r1814240, r1814326, r1814329, r1814331, r1820755 from trunk:
rand: add support for the arc4random API as an entropy source.

The arc4random API originates from OpenBSD where it supersedes random(3),
rand(3), and files in the /dev filesystem.
Use it for apr_generate_random_bytes().

Proposed by: Christian Weisgerber
Reviewed by: ylavic


rand: add support for getrandom() on Linux as an entropy source.

Use it for apr_generate_random_bytes() when available, reading from the
urandom source, and non-blocking such that the call fails with EAGAIN if
there is not enough entropy on the system (which shouldn't be the case
in userspace).


rand: follow up to r1814240.

Fall back to using SYS_getrandom syscall when it's available in linux (3.17+)
but not in glibc (2.25+, not very deployed yet).


rand: follow up to r1814240 and r1814326.

Thanks Brane, <sys/random.h> may be available w/o getrandom(), and obviously
we must not USE_GETRANDOM in this case.


rand: follow up to r1814239, r1814240 and r1814326.

The above commits added the checks for HAVE_GETRANDOM and HAVE_ARC4RANDOM first
in configure.in so to avoid breaking the AC_MESSAGE_CHECKING/RESULT line, thus
it also defined the macros unconditionally (while the others were and still are
only defined if no previous one is elected already).

Yet the top priority one should remain HAVE_EGD when --with-egd is specified,
so we now have to rearrange apr_generate_random_bytes()'s #ifdefs to preserve
that now.

While at it, let's add an #error for the "should not happen" case where
APR_HAS_RANDOM is defined but no implementation is found.


fix ifdef for arc4random

r1814239 added:
  AC_CHECK_FUNCS(arc4random_buf)

Which only defines HAVE_ARC4RANDOM_BUF


Submitted by: ylavic, covener


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.7.x@1828287 13f79535-47bb-0310-9956-ffa450edef68