Win32: Improve apr_file_read() performance on buffered files by reducing the amount of ReadFile() calls for large reads.
Previously, reading has been implemented with a loop that keeps filling in
the internal 4KB buffer and copying the data from it.  This patch reduces
the amount of syscalls for large reads by performing them with a single
syscall, if possible.  With the new approach, reads are first handled from
the buffer — however, once the buffer is empty and the remaining chunk
exceeds the capacity of the internal buffer, it will be read with a single
syscall.

Otherwise, the behavior is unchanged: we fill in the buffer up to its
capacity and serve the requested part from it.  To avoid introducing a
regression in the case when the large read happens with a single syscall,
copy the final part of the data into the internal buffer so that seeking
backwards (within the bufsize) and reading would work from the buffer.

A quick benchmark shows the significant reduction of the CPU time for
an application that uses buffered files with both small and large reads
(the large reads are approximately 128 KB-sized):

  CPU time: 27.250 s → 17.203 s
  Amount of syscalls: 1,579,587 → 139,899


* file_io/win32/readwrite.c
  (read_buffered): Reimplement the core part of this function as described
   above.  Note that the new approach no longer requires a loop, as in the
   case when the buffer is empty, we only need to make a single syscall,
   either to fill the internal buffer or to read directly into the destination
   buffer.  Rename a couple of local variables for clarity.

* test/testfile.c
  (test_empty_read_buffered,
   test_large_read_buffered,
   test_two_large_reads_buffered,
   test_small_and_large_reads_buffered,
   test_read_buffered_spanning_over_bufsize,
   test_single_byte_reads_buffered,
   test_read_buffered_seek): New tests.
  (testfile): Run the new tests.

* CHANGES: Add entry.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1828509 13f79535-47bb-0310-9956-ffa450edef68