apr_json: object keys are strings.
Make this clear in the API by requiring a usual string in apr_json_object_set()
and creating the apr_json_string_t from it.

It's also more user friendly as otherwise apr_json_string_create() is to be
used when setting an entry.

Also, axe the 'pool' arg from apr_json_array_add() since it's not needed.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1839755 13f79535-47bb-0310-9956-ffa450edef68