Merge r1856022 from trunk:
Use proc mutex pthread by default when robust[_np]

On platforms that support pshared and robust pthread mutex, this is usually
the best interprocess mutex mechanism because it's efficient, posix, not
limited and not persistent on the system when the program exits (i.e. no need
to delete it explicitely before leaving, like IPC SysV or files for instance).

Note that on older POSIX systems pthread_mutex_{setrobust,consistent}() funcs
existed with the non-posix _np() suffix, and we consider them equivalent.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.7.x@1856051 13f79535-47bb-0310-9956-ffa450edef68