Follow up to r1789258: configure to detect whether readdir() is thread-safe.
On platforms where readdir_r() is available but deprecated, readdir() is to
be used although it's not in libc_r (e.g. Linux has no libc_r).

In this case we can APR_TRY_COMPILE_NO_WARNING readdir_r() and, if it's
deprecated, define READDIR_IS_THREAD_SAFE.

With this we don't need user-defined APR_USE_READDIR{,64}_R from r1789258.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1856189 13f79535-47bb-0310-9956-ffa450edef68