Align apr_mmap()ing offset to a page boundary.  PR 65158.
This is requirement for unix systems, and otherwise can cause failures or
undefined behaviour with some filesystems.

Add apr_mmap_t->poffset on unixes to track the offset to the previous page,
relative to the user requested/actual offset. This allows ->mm and ->size to
still point to the actual data, while munmap() still deletes the full range
from the start of the page.

Tests updated accordingly.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1887060 13f79535-47bb-0310-9956-ffa450edef68