Merge r1887060 from trunk:
Align apr_mmap()ing offset to a page boundary.  PR 65158.

This is requirement for unix systems, and otherwise can cause failures or
undefined behaviour with some filesystems.

Add apr_mmap_t->poffset on unixes to track the offset to the previous page,
relative to the user requested/actual offset. This allows ->mm and ->size to
still point to the actual data, while munmap() now deletes the full range
from the start of the page.

Tests updated accordingly.


* Changes between r1887060 and this 1.7.x backport

Since the layout of struct apr_mmap_t cannot change (non opaque), the poffset
is over-allocated in mmap/unix/mmap.cmmap.c which uses this layout internally:
    struct mm_layout {
        apr_mmap_t mm;
        apr_off_t poffset;
    };

This works for all the apr_mmap_t created by apr_mmap_create() with which any
apr_mmap_*() function can then be called.

If a user creates an apr_mmap_t "by hand", with this change it cannot be passed
to apr_mmap_dup() without causing an out-of-bound read. This is hardly a new
limitation though, the refcounting which is the point of apr_mmap_dup() and
apr_mmap_delete() comes from the private function mmap_cleanup() anyway, so
there is no point in using these two functions with hand made apr_mmap_t (if
there ever was a point in the first place to create an hand made apr_mmap_t..).


Submitted by: ylavic


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.7.x@1887497 13f79535-47bb-0310-9956-ffa450edef68