apr_socket_sendv: WIN32: Follow up to r1902312: Avoid short writes.
We possibly (edge cases) can't send an entire iovec array in a single WSASend()
call because structs WSABUF and iovec are ABI incompatible.

To avoid breaking users that rely on full-write by apr_socket_sendv(), which
supposedly is guaranteed by WSASend(), repeat the call until the given iovec
array is exhausted.

There is no way to provide both full-write and atomicity guarantees for
apr_socket_sendv() on Windows, so we choose the former..

* include/apr_network_io.h:
  Document apr_socket_sendv() full-write/atomicity (non-)guarantees above the
  system ones.

* network_io/win32/sendrecv.c(apr_socket_sendv):
  Change to a loop on WSASend() when needed, taking care of its API limits
  w.r.t. the given struct iovec.



git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1904699 13f79535-47bb-0310-9956-ffa450edef68