Align the apr_memcache API with the apr_reslist API
So far apr_memcache_server_create used different types for the parameters
min, smax, max, ttl than apr_reslist_create. While the mismatch for
min, smax, max was not that much of an issue (apr_uint32_t vs. int)
as these parameters are not expected to be < 0 it was an issue for ttl
(apr_uint32_t vs. apr_interval_time_t) as this severely reduced the length
of the ttl which is in microseconds and not in seconds.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1910026 13f79535-47bb-0310-9956-ffa450edef68