Fix socket creation when APR_UNSPEC is used
When APR_UNSPEC is used, socket family is automaticaly set to APR_INET6
in apr_socket_create() and therefore if apr_sockaddr_info_get() finds out
that the given address is in IPv4 format, it fails when calling
apr_socket_connect(). So we first call apr_sockaddr_info_get() to get
info about addr family and then pass it to apr_socket_create().